refactor: get rid off barrel pattern in the Support/icons dir
