+++ src/components/Support/PartnerSection/PartnerCard.tsx
@@ -1,5 +1,6 @@
+import { PartnerCardIcon } from "../icons/PartnerCardIcon";
 import { SupportCard } from "../SupportCard";
-import { PartnerCardIcon } from "../icons";
+
 export const PartnerCard = () => (
   <SupportCard>
     <PartnerCardIcon />
