--- src/components/Support/icons/index.tsx
@@ -1,6 +1,0 @@
-export * from "./ContactIcon";
-export * from "./DonationIcon";
-export * from "./GratitudeIcon";
-export * from "./PartnerCardIcon";
-export * from "./PartnerIcon";
-export * from "./TaxIcon";
